dberror: move to utils/pkg for cross-service reuse (#696)
