+++ api/handlers/errors.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/getsentry/sentry-go"
 
-	"github.com/malbeclabs/lake/api/handlers/dberror"
+	"github.com/malbeclabs/lake/utils/pkg/dberror"
 	"github.com/malbeclabs/lake/utils/pkg/redact"
 )
