+++ api/handlers/publisher_check.go
@@ -12,7 +12,7 @@ import (
 
 	"golang.org/x/mod/semver"
 
-	"github.com/malbeclabs/lake/api/handlers/dberror"
+	"github.com/malbeclabs/lake/utils/pkg/dberror"
 	"github.com/malbeclabs/lake/api/metrics"
 )
